completed with some controllers

+++ API/app/Http/Controllers/API/JobCategoryController.php
@@ -13,28 +13,57 @@ class JobCategoryController extends Controller
         return JobCategory::all();
     }
  
-    public function show(JobCategory $job_category)
+    public function show($id)
     {
-        return $job_category;
+        $job_category = JobCategory::find($id);
+        return response()->json([
+            'status' => 200,
+            'job_category' => $job_category,
+        ],200);
     }
  
     public function store(Request $request)
     {
-        $job_category = JobCategory::create($request->all());
+        $job_category = JobCategory::create([
+            'job_name' => $request->job_name
+        ]);
  
-        return response()->json($job_category, 201);
+        if($job_category){
+            return response()->json([
+                'status' => 200,
+                'messages' => 'Job category is created successfully'
+            ],200);
+        }else{
+            return response()->json([
+                'status' => 500,
+                'messages' => 'Something went wrong'
+            ],500);
+        }
     }
  
-    public function update(Request $request, JobCategory $job_category)
+    public function update(Request $request, int $id)
     {
-        $job_category->update($request->all());
+        $job_category = JobCategory::find($id);
+        $job_category->update([
+            'job_name' => $request->job_name
+        ]);
  
-        return response()->json($job_category, 200);
     }
  
-    public function delete(JobCategory $job_category)
+    public function delete($id)
     {
-        $job_category->delete();
-        return response()->json(null, 204);
+        $job_category  = JobCategory::find($id);
+        if($job_category){
+            $job_category->delete();
+            return response()->json([
+                'status' => 200,
+                'messages' => 'Job Category is deleted successfully'
+            ],200); 
+        }else{
+            return response()->json([
+                'status' => 404,
+                'messages' => 'Delete failed'
+            ],404);
+        }
     }
 }
